schema-snapshot: move the export-data exit hooks to exportData.go
schemaSnapshotCapture.go held two unrelated groups: the reusable capture
primitives, and the export-data-specific exit-capture machinery that
only export data uses.

Move the latter -- exportDataExitSnapshotCaptured,
captureExportDataExitSnapshot, exportDataExitReason and
registerExportDataExitSnapshotHook -- into exportData.go, next to the
export-data lifecycle code that drives them. schemaSnapshotCapture.go is
now only "how to capture" and no longer needs the atexit or sync/atomic
imports at all.

Pure move: no logic change, same package throughout.
